+++ crates/core/src/gql/mod.rs
@@ -4,6 +4,7 @@ pub mod cache;
 pub mod error;
 mod ext;
 pub mod schema;
+mod tables;
 mod utils;
 
 pub use error::GqlError;
